move the logger class
